+++ CHANGELOG.md
@@ -10,6 +10,7 @@ is roughly based on [Keep a Changelog], and this project tries to adheres to
 
 - Added a by-operation iterator for alignment states (`CigletOpIter`)
 - Added `max_inc` field to `CigletSpecs` (behind `fuzzing` feature gate)
+- `DisplayErrStack` now works for `&dyn Error` (so it can be called on `Error::source`)
 
 ## [0.0.30] - 2026-06-24